Revert "Merge branch 'refactor' into documet-rag-refactor"
This reverts commit 6ec0d7d68eec1ebaf2ee96808360754549fc05dd, reversing
changes made to 7950de3373e47e8c0e2cfbab4c285aebe5702ba5.
