+++ .env.example
@@ -5,10 +5,8 @@ HTTP_PORT=3000
 # gRPC Server
 GRPC_HOST=127.0.0.1
 GRPC_PORT=50051
-
-# API keys used by both the HTTP and gRPC servers (required)
-# See keys.example.json for the file format
-VORTEXDB_KEYS_FILE=./keys.json
+# required
+GRPC_ROOT_PASSWORD=your-secure-password
 
 # Database Configuration
 # Storage: inmemory, rocksdb
