+++ crates/grpc/src/interceptors.rs
@@ -1,6 +1,3 @@
-use std::sync::Arc;
-
-use defs::{ApiKeyEntry, ApiKeyRole, ApiKeyStore};
 use tonic::{Status, service::Interceptor};
 use tracing::{Level, event};
 
@@ -8,49 +5,32 @@ use crate::constants::AUTHORIZATION_HEADER_KEY;
 
 #[derive(Clone)]
 pub struct AuthInterceptor {
-    keys: Arc<ApiKeyStore>,
-}
-
-fn extract_bearer_token<T>(req: &tonic::Request<T>) -> Option<&str> {
-    req.metadata()
-        .get(AUTHORIZATION_HEADER_KEY)?
-        .to_str()
-        .ok()?
-        .strip_prefix("Bearer ")
-        .filter(|token| !token.is_empty())
-}
-
-pub fn require_write_role<T>(req: &tonic::Request<T>) -> Result<(), Status> {
-    match req.extensions().get::<ApiKeyEntry>() {
-        Some(entry) if entry.role == ApiKeyRole::ReadWrite => Ok(()),
-        Some(_) => Err(Status::permission_denied(
-            "This api key does not have write access",
-        )),
-        None => Err(Status::unauthenticated("Invalid credentials")),
-    }
+    root_password: String,
 }
 
 impl Interceptor for AuthInterceptor {
-    fn call(&mut self, mut req: tonic::Request<()>) -> Result<tonic::Request<()>, Status> {
-        let matched = extract_bearer_token(&req)
-            .and_then(|token| self.keys.find(token))
-            .cloned();
-
-        match matched {
-            Some(entry) => {
-                req.extensions_mut().insert(entry);
-                Ok(req)
-            }
-            None => {
-                event!(Level::WARN, "Unauthorized Request");
-                Err(Status::unauthenticated("Invalid credentials"))
-            }
+    fn call(&mut self, req: tonic::Request<()>) -> Result<tonic::Request<()>, Status> {
+        let auth_token = match req.metadata().get(AUTHORIZATION_HEADER_KEY) {
+            Some(t) => t,
+            None => return Err(Status::unauthenticated("Invalid credentials")),
+        };
+        if auth_token
+            .to_str()
+            .unwrap_or_default()
+            .strip_prefix("Bearer ")
+            .unwrap_or_default()
+            == self.root_password
+        {
+            Ok(req)
+        } else {
+            event!(Level::WARN, "Unauthorized Request");
+            Err(Status::unauthenticated("Invalid credentials"))
         }
     }
 }
 
 impl AuthInterceptor {
-    pub fn new(keys: Arc<ApiKeyStore>) -> AuthInterceptor {
-        AuthInterceptor { keys }
+    pub fn new(root_password: String) -> AuthInterceptor {
+        AuthInterceptor { root_password }
     }
 }
