--- crates/http/src/constants.rs
@@ -1,1 +1,0 @@
-pub const API_KEY_HEADER: &str = "api-key";
