--- crates/index/src/flat/constants.rs
@@ -1,3 +1,0 @@
-use defs::Magic;
-
-pub const FLAT_MAGIC_BYTES: Magic = [0x00, 0x00, 0x00, 0x01];
