--- crates/index/src/hnsw/constants.rs
@@ -1,3 +1,0 @@
-use defs::Magic;
-
-pub const HNSW_MAGIC_BYTES: Magic = [0x02, 0x01, 0x03, 0x00];
