+++ crates/index/src/kd_tree/mod.rs
@@ -1,4 +1,5 @@
-pub mod constants;
+use defs::Magic;
+
 pub mod helpers;
 pub mod index;
 mod serialize;
@@ -7,6 +8,6 @@ pub mod types;
 #[cfg(test)]
 mod tests;
 
-pub use constants::KD_TREE_MAGIC_BYTES;
+pub const KD_TREE_MAGIC_BYTES: Magic = [0x00, 0x01, 0x02, 0x00];
 
 pub use index::{KDTree, KDTreeConfig};
