+++ crates/server/src/config.rs
@@ -1,42 +1,61 @@
 use api::DbConfig;
-use defs::{ApiKeyEntry, ApiKeyStore, Similarity};
+use defs::Similarity;
 use dotenv::dotenv;
 use index::{IndexType, hnsw::HnswConfig, kd_tree::KDTreeConfig};
+use snafu::prelude::*;
 use std::env;
 use std::fs;
 use std::net::SocketAddr;
 use std::path::PathBuf;
-use std::sync::Arc;
 use storage::StorageType;
 use tracing::{Level, event};
 
-use crate::constants::{
-    DEFAULT_GRPC_PORT, DEFAULT_HNSW_EF, DEFAULT_HNSW_EF_CONSTRUCTION, DEFAULT_HNSW_M,
-    DEFAULT_HNSW_MAX_LAYER, DEFAULT_HTTP_PORT, DEFAULT_KD_TREE_BALANCE_THRESHOLD,
-    DEFAULT_KD_TREE_DELETE_REBUILD_RATIO, ENV_DATA_PATH, ENV_DIMENSION, ENV_DISABLE_HTTP,
-    ENV_GRPC_HOST, ENV_GRPC_PORT, ENV_HNSW_EF, ENV_HNSW_EF_CONSTRUCTION, ENV_HNSW_M, ENV_HNSW_M0,
-    ENV_HNSW_MAX_LAYER, ENV_HTTP_HOST, ENV_HTTP_PORT, ENV_INDEX_TYPE,
-    ENV_KD_TREE_BALANCE_THRESHOLD, ENV_KD_TREE_DELETE_REBUILD_RATIO, ENV_KEYS_FILE, ENV_LOGGING,
-    ENV_SIMILARITY, ENV_STORAGE_TYPE,
-};
-pub use crate::error::{ConfigError, Result};
+const DEFAULT_HTTP_PORT: &str = "3000";
+const DEFAULT_GRPC_PORT: &str = "50051";
+const DEFAULT_HNSW_M: usize = 16;
+const DEFAULT_HNSW_MAX_LAYER: usize = 16;
+const DEFAULT_HNSW_EF_CONSTRUCTION: usize = 200;
+const DEFAULT_HNSW_EF: usize = 100;
+const DEFAULT_KD_TREE_BALANCE_THRESHOLD: f32 = 0.7;
+const DEFAULT_KD_TREE_DELETE_REBUILD_RATIO: f32 = 0.25;
 
 #[derive(Debug)]
 pub struct ServerConfig {
     pub http_addr: SocketAddr,
     pub grpc_addr: SocketAddr,
-    pub api_keys: Arc<ApiKeyStore>,
+    pub grpc_root_password: String,
     pub db_config: DbConfig,
     pub logging: bool,
     pub disable_http: bool,
 }
+
+#[derive(Debug, Snafu)]
+#[snafu(visibility(pub))]
+pub enum ConfigError {
+    #[snafu(display("Missing required environment variable: {var}"))]
+    MissingRequiredEnvVar { var: String },
+
+    #[snafu(display("Invalid dimension value"))]
+    InvalidDimension,
+
+    #[snafu(display("Invalid data path: {source}"))]
+    InvalidDataPath { source: std::io::Error },
+
+    #[snafu(display("IO error: {source}"))]
+    IoError { source: std::io::Error },
+
+    #[snafu(display("Invalid address: {addr}"))]
+    InvalidAddress { addr: String },
+}
+
+pub type Result<T, E = ConfigError> = std::result::Result<T, E>;
 
 impl ServerConfig {
     pub fn load_config() -> Result<Self> {
         dotenv().ok();
 
         // HTTP server configuration
-        let http_host = env::var(ENV_HTTP_HOST)
+        let http_host = env::var("HTTP_HOST")
             .inspect_err(|_| {
                 event!(
                     Level::WARN,
@@ -45,7 +64,7 @@ impl ServerConfig {
             })
             .unwrap_or_else(|_| "127.0.0.1".to_string());
 
-        let http_port = env::var(ENV_HTTP_PORT)
+        let http_port = env::var("HTTP_PORT")
             .inspect_err(|_| {
                 event!(
                     Level::WARN,
@@ -63,7 +82,7 @@ impl ServerConfig {
                 })?;
 
         // gRPC server configuration
-        let grpc_host = env::var(ENV_GRPC_HOST)
+        let grpc_host = env::var("GRPC_HOST")
             .inspect_err(|_| {
                 event!(
                     Level::WARN,
@@ -72,7 +91,7 @@ impl ServerConfig {
             })
             .unwrap_or_else(|_| "127.0.0.1".to_string());
 
-        let grpc_port = env::var(ENV_GRPC_PORT)
+        let grpc_port = env::var("GRPC_PORT")
             .inspect_err(|_| {
                 event!(
                     Level::WARN,
@@ -89,14 +108,14 @@ impl ServerConfig {
                     addr: format!("{}:{}", grpc_host, grpc_port),
                 })?;
 
-        let keys_file_path =
-            env::var(ENV_KEYS_FILE).map_err(|_| ConfigError::MissingRequiredEnvVar {
-                var: ENV_KEYS_FILE.to_string(),
+        // gRPC root password (required)
+        let grpc_root_password =
+            env::var("GRPC_ROOT_PASSWORD").map_err(|_| ConfigError::MissingRequiredEnvVar {
+                var: "GRPC_ROOT_PASSWORD".to_string(),
             })?;
-        let api_keys = Arc::new(load_keys_file(&keys_file_path)?);
 
         // Storage type
-        let storage_type_str = env::var(ENV_STORAGE_TYPE)
+        let storage_type_str = env::var("STORAGE_TYPE")
             .inspect_err(|_| {
                 event!(
                     Level::WARN,
@@ -112,7 +131,7 @@ impl ServerConfig {
         };
 
         // Index type
-        let index_type_str = env::var(ENV_INDEX_TYPE)
+        let index_type_str = env::var("INDEX_TYPE")
             .inspect_err(|_| {
                 event!(Level::WARN, "INDEX_TYPE not defined, defaulting to flat");
             })
@@ -127,15 +146,15 @@ impl ServerConfig {
         };
 
         // Dimension (required)
-        let dimension: usize = env::var(ENV_DIMENSION)
+        let dimension: usize = env::var("DIMENSION")
             .map_err(|_| ConfigError::MissingRequiredEnvVar {
-                var: ENV_DIMENSION.to_string(),
+                var: "DIMENSION".to_string(),
             })?
             .parse()
             .map_err(|_| ConfigError::InvalidDimension)?;
 
         // Data path
-        let data_path: PathBuf = if let Ok(data_path_str) = env::var(ENV_DATA_PATH) {
+        let data_path: PathBuf = if let Ok(data_path_str) = env::var("DATA_PATH") {
             let path = PathBuf::from(data_path_str);
             fs::create_dir_all(&path).map_err(|e| ConfigError::InvalidDataPath { source: e })?;
             path
@@ -151,19 +170,19 @@ impl ServerConfig {
         };
 
         // Logging
-        let logging = env::var(ENV_LOGGING)
+        let logging = env::var("LOGGING")
             .unwrap_or_else(|_| "true".to_string())
             .parse()
             .unwrap_or(true);
 
         // HTTP server disable flag (default to false, set to true to run only gRPC)
-        let disable_http = env::var(ENV_DISABLE_HTTP)
+        let disable_http = env::var("DISABLE_HTTP")
             .unwrap_or_else(|_| "false".to_string())
             .parse()
             .unwrap_or(false);
 
         // Similarity metric
-        let similarity: Similarity = match env::var(ENV_SIMILARITY) {
+        let similarity: Similarity = match env::var("SIMILARITY") {
             Ok(val) => match val.to_lowercase().as_str() {
                 "cosine" => Similarity::Cosine,
                 "euclidean" => Similarity::Euclidean,
@@ -184,21 +203,21 @@ impl ServerConfig {
             }
         };
 
-        let hnsw_m = load_usize_env(ENV_HNSW_M, DEFAULT_HNSW_M);
+        let hnsw_m = load_usize_env("HNSW_M", DEFAULT_HNSW_M);
         let hnsw_config = HnswConfig {
             max_connections: hnsw_m,
-            max_connections_0: load_usize_env(ENV_HNSW_M0, 2 * hnsw_m),
-            max_layer: load_usize_env(ENV_HNSW_MAX_LAYER, DEFAULT_HNSW_MAX_LAYER),
-            ef_construction: load_usize_env(ENV_HNSW_EF_CONSTRUCTION, DEFAULT_HNSW_EF_CONSTRUCTION),
-            ef: load_usize_env(ENV_HNSW_EF, DEFAULT_HNSW_EF),
+            max_connections_0: load_usize_env("HNSW_M0", 2 * hnsw_m),
+            max_layer: load_usize_env("HNSW_MAX_LAYER", DEFAULT_HNSW_MAX_LAYER),
+            ef_construction: load_usize_env("HNSW_EF_CONSTRUCTION", DEFAULT_HNSW_EF_CONSTRUCTION),
+            ef: load_usize_env("HNSW_EF", DEFAULT_HNSW_EF),
         };
         let kd_tree_config = KDTreeConfig {
             balance_threshold: load_f32_env(
-                ENV_KD_TREE_BALANCE_THRESHOLD,
+                "KD_TREE_BALANCE_THRESHOLD",
                 DEFAULT_KD_TREE_BALANCE_THRESHOLD,
             ),
             delete_rebuild_ratio: load_f32_env(
-                ENV_KD_TREE_DELETE_REBUILD_RATIO,
+                "KD_TREE_DELETE_REBUILD_RATIO",
                 DEFAULT_KD_TREE_DELETE_REBUILD_RATIO,
             ),
         };
@@ -216,51 +235,12 @@ impl ServerConfig {
         Ok(ServerConfig {
             http_addr,
             grpc_addr,
-            api_keys,
+            grpc_root_password,
             db_config,
             logging,
             disable_http,
         })
     }
-}
-
-#[derive(serde::Deserialize)]
-struct KeysFile {
-    keys: Vec<ApiKeyEntry>,
-}
-
-fn load_keys_file(path: &str) -> Result<ApiKeyStore> {
-    let contents = fs::read_to_string(path).map_err(|source| ConfigError::KeysFileRead {
-        path: path.to_string(),
-        source,
-    })?;
-    let parsed: KeysFile =
-        serde_json::from_str(&contents).map_err(|source| ConfigError::KeysFileParse {
-            path: path.to_string(),
-            source,
-        })?;
-
-    if parsed.keys.is_empty() {
-        return Err(ConfigError::KeysFileEmpty {
-            path: path.to_string(),
-        });
-    }
-
-    if parsed.keys.iter().any(|entry| entry.key.is_empty()) {
-        return Err(ConfigError::EmptyApiKey {
-            path: path.to_string(),
-        });
-    }
-
-    let store = ApiKeyStore::new(parsed.keys);
-    if let Some(key) = store.duplicate_key() {
-        return Err(ConfigError::DuplicateApiKey {
-            path: path.to_string(),
-            key: key.to_string(),
-        });
-    }
-
-    Ok(store)
 }
 
 fn load_usize_env(name: &str, default: usize) -> usize {
