--- crates/server/src/constants.rs
@@ -1,28 +1,0 @@
-pub const ENV_HTTP_HOST: &str = "HTTP_HOST";
-pub const ENV_HTTP_PORT: &str = "HTTP_PORT";
-pub const ENV_GRPC_HOST: &str = "GRPC_HOST";
-pub const ENV_GRPC_PORT: &str = "GRPC_PORT";
-pub const ENV_KEYS_FILE: &str = "VORTEXDB_KEYS_FILE";
-pub const ENV_STORAGE_TYPE: &str = "STORAGE_TYPE";
-pub const ENV_INDEX_TYPE: &str = "INDEX_TYPE";
-pub const ENV_DIMENSION: &str = "DIMENSION";
-pub const ENV_DATA_PATH: &str = "DATA_PATH";
-pub const ENV_LOGGING: &str = "LOGGING";
-pub const ENV_DISABLE_HTTP: &str = "DISABLE_HTTP";
-pub const ENV_SIMILARITY: &str = "SIMILARITY";
-pub const ENV_HNSW_M: &str = "HNSW_M";
-pub const ENV_HNSW_M0: &str = "HNSW_M0";
-pub const ENV_HNSW_MAX_LAYER: &str = "HNSW_MAX_LAYER";
-pub const ENV_HNSW_EF_CONSTRUCTION: &str = "HNSW_EF_CONSTRUCTION";
-pub const ENV_HNSW_EF: &str = "HNSW_EF";
-pub const ENV_KD_TREE_BALANCE_THRESHOLD: &str = "KD_TREE_BALANCE_THRESHOLD";
-pub const ENV_KD_TREE_DELETE_REBUILD_RATIO: &str = "KD_TREE_DELETE_REBUILD_RATIO";
-
-pub const DEFAULT_HTTP_PORT: &str = "3000";
-pub const DEFAULT_GRPC_PORT: &str = "50051";
-pub const DEFAULT_HNSW_M: usize = 16;
-pub const DEFAULT_HNSW_MAX_LAYER: usize = 16;
-pub const DEFAULT_HNSW_EF_CONSTRUCTION: usize = 200;
-pub const DEFAULT_HNSW_EF: usize = 100;
-pub const DEFAULT_KD_TREE_BALANCE_THRESHOLD: f32 = 0.7;
-pub const DEFAULT_KD_TREE_DELETE_REBUILD_RATIO: f32 = 0.25;
