+++ crates/server/src/main.rs
@@ -1,6 +1,4 @@
 mod config;
-mod constants;
-mod error;
 
 use std::sync::Arc;
 
@@ -36,10 +34,7 @@ async fn main() -> Result<(), BoxError> {
     let http_handle = if !config.disable_http {
         let db = Arc::clone(&shared_db);
         let addr = config.http_addr;
-        let keys = Arc::clone(&config.api_keys);
-        Some(tokio::spawn(async move {
-            run_http_server(db, addr, keys).await
-        }))
+        Some(tokio::spawn(async move { run_http_server(db, addr).await }))
     } else {
         info!("HTTP server is disabled");
         None
@@ -49,9 +44,9 @@ async fn main() -> Result<(), BoxError> {
     let grpc_handle = {
         let db = Arc::clone(&shared_db);
         let addr = config.grpc_addr;
-        let keys = Arc::clone(&config.api_keys);
+        let password = config.grpc_root_password;
         let logging = config.logging;
-        tokio::spawn(async move { run_grpc_server(db, addr, keys, logging).await })
+        tokio::spawn(async move { run_grpc_server(db, addr, password, logging).await })
     };
 
     if let Some(http) = http_handle {
