+++ crates/snapshot/src/registry/constants.rs
@@ -1,3 +1,1 @@
 pub const LOCAL_REGISTRY_LOCKFILE: &str = "LOCKFILE";
-pub const INFINITY_LIMIT: usize = 100000;
-pub const NO_OFFSET: usize = 0;
