--- crates/storage/src/rocks_db/constants.rs
@@ -1,2 +1,0 @@
-pub const ROCKSDB_CHECKPOINT_FILENAME_MARKER: &str = "rocksdb";
-pub const ROCKSDB_CHECKPOINT_EXTENSION: &str = "tar.gz";
