+++ docker-compose.yml
@@ -20,7 +20,7 @@ services:
       DISABLE_HTTP: ${DISABLE_HTTP:-false}
 
       # Required
-      VORTEXDB_KEYS_FILE: /run/secrets/keys.json
+      GRPC_ROOT_PASSWORD: ${GRPC_ROOT_PASSWORD}
       DIMENSION: ${DIMENSION}
 
       DATA_PATH: /data
@@ -31,7 +31,6 @@ services:
 
     volumes:
       - vortexdb_data:/data
-      - ./keys.json:/run/secrets/keys.json:ro
 
 volumes:
   vortexdb_data:
