+++ docs/concepts/architecture.mdx
@@ -147,7 +147,7 @@ VortexDB is configured via environment variables:
 
 | Variable | Required | Default | Description |
 |----------|----------|---------|-------------|
-| `VORTEXDB_KEYS_FILE` | Yes | - | Path to a JSON file of API keys shared by the HTTP and gRPC servers |
+| `GRPC_ROOT_PASSWORD` | Yes | - | Authentication password for gRPC |
 | `DIMENSION` | Yes | - | Vector dimensionality |
 | `DATA_PATH` | No | system temp dir | Directory for persistent storage |
 | `HTTP_HOST` | No | `127.0.0.1` | HTTP server bind address |
