+++ docs/getting-started/installation.mdx
@@ -37,7 +37,7 @@ Edit the `.env` file with your settings:
 
 ```bash
 # Required settings
-VORTEXDB_KEYS_FILE=./keys.json  # see keys.example.json for the format
+GRPC_ROOT_PASSWORD=your-secure-password
 DIMENSION=384  # Vector dimension (e.g., 384 for MiniLM embeddings)
 DATA_PATH=/data
